🔨 Refactor: Mypage components/fix session errors
